requirements.txt
flask
pillow
pillow-heif
werkzeug
